Reworked the collision resolving.
The algorithm uses a map with a list of collided objects as value, now.
That allows me to resolve more than one objects at the same time instead
of one by another in a kind of sequence.

The problem was that it is simply possible that an object collide with
more than one object at the same time and it wasn't possible to handle
it properly.

In addition, i removed some useless legacy classes, that i wont use
probably.