Fixed some Collisionsbugs. Removed actual Quadtree cause it didn't work.
First bug was that after an object collided, it was stuck. He object
couldn't recalculate the velocity cause of an if-statement. The solution
is to separate the calculation of the velocity from the update and
collision resolving and put it in a correct order in the game loop.

Second bug was in the Quadtree. The retrieve-Method that i stole didn't
work.

+ Removed unnecessary methods and classes.